requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.4.0
sse-starlette>=1.8.0
ollama>=0.1.6
duckduckgo-search>=4.4.0
beautifulsoup4>=4.12.0
langgraph>=0.0.26
langchain-core>=0.1.30
aiohttp>=3.9.0
pytest>=7.4.0
httpx>=0.27.0
